mknod: completions
