+++ src/uu/mknod/Cargo.toml
@@ -21,6 +21,9 @@ libc = "^0.2.42"
 uucore = { version=">=0.0.8", package="uucore", path="../../uucore", features=["mode"] }
 uucore_procs = { version=">=0.0.5", package="uucore_procs", path="../../uucore_procs" }
 
+[build-dependencies]
+clap = "2.33"
+
 [[bin]]
 name = "mknod"
 path = "src/main.rs"
