+++ src/uu/mknod/build.rs
@@ -1,0 +1,2 @@
+include!("src/app.rs");
+include!("../../build_completions.rs");
