+++ src/uu/mknod/src/mknod.rs
@@ -12,33 +12,15 @@ extern crate uucore;
 
 use std::ffi::CString;
 
-use clap::{crate_version, App, Arg, ArgMatches};
+use clap::ArgMatches;
 use libc::{dev_t, mode_t};
 use libc::{S_IFBLK, S_IFCHR, S_IFIFO, S_IRGRP, S_IROTH, S_IRUSR, S_IWGRP, S_IWOTH, S_IWUSR};
 
 use uucore::InvalidEncodingHandling;
 
-static NAME: &str = "mknod";
-static ABOUT: &str = "Create the special file NAME of the given TYPE.";
-static USAGE: &str = "mknod [OPTION]... NAME TYPE [MAJOR MINOR]";
-static LONG_HELP: &str = "Mandatory arguments to long options are mandatory for short options too.
--m, --mode=MODE    set file permission bits to MODE, not a=rw - umask
---help     display this help and exit
---version  output version information and exit
+use crate::app::get_app;
 
-Both MAJOR and MINOR must be specified when TYPE is b, c, or u, and they
-must be omitted when TYPE is p.  If MAJOR or MINOR begins with 0x or 0X,
-it is interpreted as hexadecimal; otherwise, if it begins with 0, as octal;
-otherwise, as decimal.  TYPE may be:
-
-b      create a block (buffered) special file
-c, u   create a character (unbuffered) special file
-p      create a FIFO
-
-NOTE: your shell may have its own version of mknod, which usually supersedes
-the version described here.  Please refer to your shell's documentation
-for details about the options it supports.
-";
+mod app;
 
 const MODE_RW_UGO: mode_t = S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH;
 
@@ -72,7 +54,7 @@ fn _mknod(file_name: &str, mode: mode_t, dev: dev_t) -> i32 {
         }
 
         if errno == -1 {
-            let c_str = CString::new(NAME).expect("Failed to convert to CString");
+            let c_str = CString::new(executable!()).expect("Failed to convert to CString");
             // shows the error from the mknod syscall
             libc::perror(c_str.as_ptr());
         }
@@ -89,48 +71,7 @@ pub fn uumain(args: impl uucore::Args) -> i32 {
     // opts.optflag("Z", "", "set the SELinux security context to default type");
     // opts.optopt("", "context", "like -Z, or if CTX is specified then set the SELinux or SMACK security context to CTX");
 
-    let matches = App::new(executable!())
-        .version(crate_version!())
-        .usage(USAGE)
-        .after_help(LONG_HELP)
-        .about(ABOUT)
-        .arg(
-            Arg::with_name("mode")
-                .short("m")
-                .long("mode")
-                .value_name("MODE")
-                .help("set file permission bits to MODE, not a=rw - umask"),
-        )
-        .arg(
-            Arg::with_name("name")
-                .value_name("NAME")
-                .help("name of the new file")
-                .required(true)
-                .index(1),
-        )
-        .arg(
-            Arg::with_name("type")
-                .value_name("TYPE")
-                .help("type of the new file (b, c, u or p)")
-                .required(true)
-                .validator(valid_type)
-                .index(2),
-        )
-        .arg(
-            Arg::with_name("major")
-                .value_name("MAJOR")
-                .help("major file type")
-                .validator(valid_u64)
-                .index(3),
-        )
-        .arg(
-            Arg::with_name("minor")
-                .value_name("MINOR")
-                .help("minor file type")
-                .validator(valid_u64)
-                .index(4),
-        )
-        .get_matches_from(args);
+    let matches = get_app(executable!()).get_matches_from(args);
 
     let mode = match get_mode(&matches) {
         Ok(mode) => mode,
@@ -154,7 +95,7 @@ pub fn uumain(args: impl uucore::Args) -> i32 {
     if ch == 'p' {
         if matches.is_present("major") || matches.is_present("minor") {
             eprintln!("Fifos do not have major and minor device numbers.");
-            eprintln!("Try '{} --help' for more information.", NAME);
+            eprintln!("Try '{} --help' for more information.", executable!());
             1
         } else {
             _mknod(file_name, S_IFIFO | mode, 0)
@@ -163,7 +104,7 @@ pub fn uumain(args: impl uucore::Args) -> i32 {
         match (matches.value_of("major"), matches.value_of("minor")) {
             (None, None) | (_, None) | (None, _) => {
                 eprintln!("Special files require major and minor device numbers.");
-                eprintln!("Try '{} --help' for more information.", NAME);
+                eprintln!("Try '{} --help' for more information.", executable!());
                 1
             }
             (Some(major), Some(minor)) => {
@@ -198,23 +139,4 @@ fn get_mode(matches: &ArgMatches) -> Result<mode_t, String> {
                 }
             }),
     }
-}
-
-fn valid_type(tpe: String) -> Result<(), String> {
-    // Only check the first character, to allow mnemonic usage like
-    // 'mknod /dev/rst0 character 18 0'.
-    tpe.chars()
-        .next()
-        .ok_or_else(|| "missing device type".to_string())
-        .and_then(|first_char| {
-            if vec!['b', 'c', 'u', 'p'].contains(&first_char) {
-                Ok(())
-            } else {
-                Err(format!("invalid device type ‘{}’", tpe))
-            }
-        })
-}
-
-fn valid_u64(num: String) -> Result<(), String> {
-    num.parse::<u64>().map(|_| ()).map_err(|_| num)
 }
